+++ src/wymeditor/plugins/table/jquery.wymeditor.table.js
@@ -138,22 +138,38 @@ TableEditor.prototype.bindEvents = function () {
     jQuery(wym._box).find(
         tableEditor._options.sAddRowButtonSelector
     ).click(function () {
-        return tableEditor.addRow(wym.selectedContainer());
+        var cont = wym.selectedContainer();
+        if (cont) {
+            tableEditor.addRow(cont);
+        }
+        return false;
     });
     jQuery(wym._box).find(
         tableEditor._options.sRemoveRowButtonSelector
     ).click(function () {
-        return tableEditor.removeRow(wym.selectedContainer());
+        var cont = wym.selectedContainer();
+        if (cont) {
+            tableEditor.removeRow(cont);
+        }
+        return false;
     });
     jQuery(wym._box).find(
         tableEditor._options.sAddColumnButtonSelector
     ).click(function () {
-        return tableEditor.addColumn(wym.selectedContainer());
+        var cont = wym.selectedContainer();
+        if (cont) {
+            tableEditor.addColumn(cont);
+        }
+        return false;
     });
     jQuery(wym._box).find(
         tableEditor._options.sRemoveColumnButtonSelector
     ).click(function () {
-        return tableEditor.removeColumn(wym.selectedContainer());
+        var cont = wym.selectedContainer();
+        if (cont) {
+            tableEditor.removeColumn(cont);
+        }
+        return false;
     });
 
     // Handle tab clicks
@@ -394,13 +410,13 @@ TableEditor.prototype.mergeRow = function () {
     // Just use the td and th nodes
     cells = jQuery(nodes).filter('td,th');
     if (cells.length === 0) {
-        return false;
+        return;
     }
 
     // If the selection is across multiple tables, don't merge
     rootTr = tableEditor.getCommonParentTr(cells);
     if (rootTr === null) {
-        return false;
+        return;
     }
 
     mergeCell = cells[0];
@@ -516,7 +532,7 @@ TableEditor.prototype.addRow = function (elmnt) {
         i;
 
     if (tr === null) {
-        return false;
+        return;
     }
 
     numColumns = tableEditor.getNumColumns(tr);
@@ -528,7 +544,6 @@ TableEditor.prototype.addRow = function (elmnt) {
     jQuery(tr).after('<tr>' + tdHtml + '</tr>');
 
     wym.registerModification();
-    return false;
 };
 
 /**
@@ -563,7 +578,7 @@ TableEditor.prototype.removeRow = function (elmnt) {
         table;
 
     if (tr === null) {
-        return false;
+        return;
     }
     table = wym.findUp(elmnt, 'table');
     var $tr = jQuery(tr);
@@ -586,7 +601,6 @@ TableEditor.prototype.removeRow = function (elmnt) {
     tableEditor.removeEmptyTable(table);
 
     wym.registerModification();
-    return false;
 };
 
 /**
@@ -606,7 +620,7 @@ TableEditor.prototype.addColumn = function (elmnt) {
         insertionElement;
 
     if (td === null) {
-        return false;
+        return;
     }
     prevTds = jQuery(td).prevAll();
     tdIndex = prevTds.length;
@@ -623,7 +637,6 @@ TableEditor.prototype.addColumn = function (elmnt) {
     });
 
     wym.registerModification();
-    return false;
 };
 
 /**
@@ -639,7 +652,7 @@ TableEditor.prototype.removeColumn = function (elmnt) {
         tdIndex,
         tr;
     if (td === null) {
-        return false;
+        return;
     }
     table = wym.findUp(elmnt, 'table');
     prevTds = jQuery(td).prevAll();
@@ -668,7 +681,6 @@ TableEditor.prototype.removeColumn = function (elmnt) {
     tableEditor.removeEmptyTable(table);
 
     wym.registerModification();
-    return false;
 };
 
 /**